% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AICquasiPoisson.R
\name{.AICquasiPoisson}
\alias{.AICquasiPoisson}
\title{AIC for Quasi-Poisson glm model}
\usage{
.AICquasiPoisson(fitObj)
}
\arguments{
\item{fitObj}{a fitted model object}
}
\value{
numeric AIC for Quasi-Poisson glm model
}
\description{
AIC for Quasi-Poisson glm model dpois
}
\keyword{internal}
