% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beta_bin_meth.R
\name{beta_bin_meth}
\alias{beta_bin_meth}
\alias{beta_bin_meth,matrix_OR_data.frame-method}
\alias{beta_bin_meth,GRanges-method}
\alias{beta_bin_meth,GRangesList-method}
\title{Beta-binomial Posterior Methylation Levels}
\usage{
beta_bin_meth(x, ...)

\S4method{beta_bin_meth}{matrix_OR_data.frame}(
  x,
  init.pars = NULL,
  via.optim = TRUE,
  loss.fun = c("linear", "huber", "smooth", "cauchy", "arctg"),
  verbose = TRUE,
  ...
)

\S4method{beta_bin_meth}{GRanges}(
  x,
  init.pars = NULL,
  via.optim = TRUE,
  loss.fun = c("linear", "huber", "smooth", "cauchy", "arctg"),
  verbose = TRUE,
  ...
)

\S4method{beta_bin_meth}{GRangesList}(
  x,
  init.pars = NULL,
  via.optim = TRUE,
  loss.fun = c("linear", "huber", "smooth", "cauchy", "arctg"),
  num.cores = multicoreWorkers(),
  tasks = 0L,
  verbose = TRUE,
  ...
)
}
\arguments{
\item{x}{A \code{\link[GenomicRanges]{GRanges-class}} carrying a matrix of
counts on its metacolumn, with the counts of methylated (mC) and
unmethylated (uC) cytosines for at least 10 or more cytosine sites.
Alternatively, it can be just the 'matrix' or 'data.frame' of counts.}

\item{init.pars}{initial parameter values. Defaults is NULL and an initial
guess is estimated using \code{\link[stats]{optim}} function. If the initial
guessing fails initial parameter values are to alpha = 1 &
beta = 1, which imply the parsimony pseudo-counts greater than zero.}

\item{via.optim}{Logical. Whether to estimate beta distribution parameters
via \code{\link[stats]{optim}} or \code{\link[minpack.lm]{nls.lm}}. If any of
this approaches fail then parameters used \emph{init.pars} will be returned.}

\item{loss.fun}{Loss function(s) used in the estimation of the best fitted
model to beta distribution (only applied when Bayesian=TRUE; see
\href{https://en.wikipedia.org/wiki/Loss_function}{(Loss function)}). This
fitting uses the approach followed in in the R package
\href{https://github.com/genomaths/usefr}{usefr}. After
\eqn{z =  1/2 * sum((f(x) - y)^2)} we have:
\enumerate{
\item "linear": linear function which gives a standard least squares:
\eqn{loss(z) = z}.
\item "huber": Huber loss,
\eqn{loss(z) = ifelse(z \leq 1, z, sqrt(z) -1)}.
\item "smooth": Smooth approximation to the sum of residues absolute
values: \eqn{loss(z) = 2*(sqrt(z + 1) - 1)}.
\item "cauchy": Cauchy loss: \eqn{loss(z) = log(z + 1)}.
\item "arctg": arc-tangent loss function: \eqn{loss(x) = atan(z)}.
}
Loss 'linear' function works well for most of the methylation datasets with
acceptable quality.}
}
\description{
This a function is addressed to estimate the posterior
probabilities of methylation levels, assuming that the methylation levels
follows Beta distribution and taking abventage that ehe Beta distribution is
a conjugate prior for Binomial distribution.
}
\author{
Robersy Sanchez (\url{https://genomaths.com})
}
