% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boltzman_factor.R
\name{boltzman_factor}
\alias{boltzman_factor}
\alias{boltzman_factor,missingORNULL-method}
\alias{boltzman_factor,cdfMODEL-method}
\alias{boltzman_factor,cdfMODELlist-method}
\alias{boltzman_factor,ProbDistrList-method}
\title{Boltzmann's Factors}
\usage{
boltzman_factor(model = NULL, pars, ...)

\S4method{boltzman_factor}{missingORNULL}(model, pars, only.sum = TRUE)

\S4method{boltzman_factor}{cdfMODEL}(model, only.sum = TRUE)

\S4method{boltzman_factor}{cdfMODELlist}(model, only.sum = TRUE)

\S4method{boltzman_factor}{ProbDistrList}(model, only.sum = TRUE)
}
\arguments{
\item{model}{An object from any of the classes created in MethylIT
pipeline: \emph{\strong{cdfMODEL}}, \emph{\strong{cdfMODELlist}}, or
\emph{\strong{ProbDistrList}}. If given, then the parameter values are taken
from the model.}

\item{pars}{Optional. A numerical vector containing the model parameter
values in the given in order: \strong{\emph{alpha, scale}}, and
\strong{\emph{delta}}.}

\item{only.sum}{If only.sum = TRUE, then only the sum of Boltzmann's factors
is returned.}
}
\value{
Boltzmann's factors \eqn{exp(-abs(S)/k_B)}, \eqn{exp(-\nu)}, and
\eqn{exp(-abs(S)/k_B) + exp(-\nu)}, and the expected values \eqn{\nu} and
the variance \eqn{\sigma = \theta^2 Gamma(\psi + 2/\alpha)/\Gamma(\psi)}.
If only.sum = TRUE (default), the only the sum of Boltzmann's factors is
returned.
}
\description{
This function computes the Boltzmann's factors as function of
addimensional Gibb entropy \strong{\eqn{S/k_B}}
(see \code{\link{gibb_entropy}}) and the expected value \strong{\eqn{\nu}}
of the generalized gamma (GG) distribution (with location parameter
\eqn{\mu = 0}):

\deqn{exp(-y^\alpha) * \alpha*y^(\alpha*\delta - 1)/(scale*\gamma(\delta))}

(see \href{https://goo.gl/1n4kpW}{(Wikipedia)})

The expected value \strong{\eqn{\nu}} is given as:

\deqn{\nu = \theta * \Gamma(\psi + 1/\alpha)/\Gamma(\psi)}

(\eqn{\\theta = scale})
}
\details{
The Boltzmann's factor of Gibb entropy \strong{\eqn{S/k_B}} is
given as:

\deqn{exp(-abs(S)/k_B)}

While the Boltzmann's factor of \strong{\eqn{\nu}} is calculated as:

\deqn{exp(-\eqn{\nu})}
}
\examples{
## Loading the probability distribution models
data(gof, "MethylIT")
## Gibb entropy in J * (K * mol)^-1)
gibb_entropy(gof)
## Shannon entropy
gibb_entropy(gof, R = 1, log.base = 2)
}
\seealso{
\code{\link{gibb_entropy}}
}
