% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap2x2.R
\name{bootstrap2x2}
\alias{bootstrap2x2}
\title{bootstrap2x2}
\usage{
bootstrap2x2(x, stat = "rmst", num.permut = 100)
}
\arguments{
\item{x}{A numerical matrix corresponding to cross tabulation (2x2) table
(contingency table).}

\item{stat}{Statistic to be used in the testing: 'RMST' or 'HD'}

\item{num.permut}{Number of permutations.}
}
\value{
A p-value probability
}
\description{
Parametric Bootstrap of 2x2 Contingence independence test. The
    goodness of fit statistic is the root-mean-square statistic (RMST) or
    Hellinger divergence, as proposed by Perkins et al. [1, 2]. Hellinger
    divergence (HD) is computed as proposed in [3].
}
\examples{
    set.seed(123)
    TeaTasting = matrix(c(8, 350, 2, 20), nrow = 2,
                        dimnames = list(Guess = c("Milk", "Tea"),
                        Truth = c("Milk", "Tea")))
    ## Small num.permut for test's speed sake
    bootstrap2x2( TeaTasting, stat = "all", num.permut = 100 )
}
\references{
1. Perkins W, Tygert M, Ward R. and Classical Exact Tests Often
        Wildly Misreport Significance; the Remedy Lies in Computers
        [Internet]. Uploaded to ArXiv. 2011. Report No.: arXiv:1108.4126v2.
    2. Perkins, W., Tygert, M. & Ward, R. Computing the confidence levels
        for a root-mean square test of goodness-of-fit. 217, 9072-9084
        (2011).
    3. Basu, A., Mandal, A. & Pardo, L. Hypothesis testing for two discrete
        populations based on the Hellinger distance. Stat. Probab. Lett. 80,
        206-214 (2010).
}
