% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coef.R
\name{coef.cdfMODEL}
\alias{coef.cdfMODEL}
\alias{coef}
\alias{coef.cdfMODELlist}
\alias{coef.ProbDistrList}
\title{Extract Model Coefficients}
\usage{
\method{coef}{cdfMODEL}(object)

\method{coef}{cdfMODELlist}(object)

\method{coef}{ProbDistrList}(object)
}
\arguments{
\item{object}{an object for which the extraction of model coefficients is
meaningful.}
}
\description{
This function extend \code{\link[stats]{coef}} permiting the
coefficient extraction from the non-linear models estimated in
\strong{MethylIT}.
}
