% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colorBar.R
\name{.colorBar}
\alias{.colorBar}
\title{Color bar for heatmaps}
\usage{
.colorBar(z, zlim, col = heat.colors(12), breaks, horiz = TRUE,
  ylim = NULL, xlim = NULL, lwd = 0.5, cex.axis = 1, ...)
}
\arguments{
\item{z}{matrix of values used to visualize the heatmap}

\item{zlim}{Limit for the numerical(color) scale, which must be consistent
with parameter 'break'.}

\item{col}{Palette of color to use}

\item{breaks}{Numerical vector with the breaks used to produce the heatmap}

\item{horiz}{Whether to the color bar will be horizontal(= TRUE) or
vertical(= FALSE)}

\item{ylim}{User defined limits for y-axis. Depending on the orientation, x-
or y-limits may be defined that are different from the z-limits and will
reduce the range of colors displayed.}

\item{xlim}{User defined limits for x-axis}

\item{lwd}{Line width.}

\item{cex.axis}{Cex values for color bar axis labels.}

\item{...}{Additional parameter to pass to 'par' R function}
}
\value{
Image with color bar
}
\description{
This function is for internal use. It creates a color scaled
    bar for heatmap.
}
