% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computeSizeFactors.R
\name{.computeSizeFactors}
\alias{.computeSizeFactors}
\title{Estimate size factors}
\usage{
.computeSizeFactors(dss)
}
\arguments{
\item{dss}{DESeqDataSet object}
}
\value{
Estimation of the size factors
}
\description{
Size factor estimation provided by DEseq
}
\details{
A comment about size factor estimation by DEseq based on Simon
    Anders (Heidelberg, Germany) work to estimate the library size, simply
    taking the total number of (mapped or unmapped) reads is, in our
    experience, not a good idea. Sometimes, a few very strongly expressed
    genes are differentially expressed, and as they make up a good part of
    the total counts, they skew this number. After you divide by total
    counts, these few strongly expressed genes become equal, and the whole
    rest looks differentially expressed.

    The following simple alternative works much better: - Construct a
    "reference sample" by taking, for each gene, the geometric mean of the
    counts in all samples. - To get the sequencing depth of a sample
    relative to the reference, calculate for each gene the quotient of the
    counts in your sample divided by the counts of the reference sample. Now
    you have, for each gene, an estimate of the depth ratio. - Simply take
    the median of all the quotients to get the relative depth of the
    library. This is what the 'estimateSizeFactors' function of our DESeq
    package does.
}
\examples{
    dds <- makeExampleDESeqDataSet(n = 1000, m = 4)
    MethylIT:::.computeSizeFactors(dds)

}
\references{
http://seqanswers.com/forums/showpost.php?p=16468&postcount=13
}
