% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_cutpoint.R
\docType{data}
\name{cutpoint}
\alias{cutpoint}
\title{Cutpoint of the 'PS' simulated dataset used in the examples}
\format{
A list with the cutpoint value and classification performance. The
cutpoint values was obtained with function
\code{\link{estimateCutPoint}}.
}
\usage{
cutpoint
}
\description{
Each sample individual in the 'PS' included 10000 cytosine positions
}
\keyword{datasets}
