% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimateJDiv.R
\name{estimateJDiv}
\alias{estimateJDiv}
\title{J Information Divergence of Methylation Levels}
\usage{
estimateJDiv(p, logbase = 2)
}
\arguments{
\item{p}{A numerical vector of the methylation levels p = c(p1, p2) from
individuals 1 and 2.}

\item{logbase}{Logarithm base used to compute the JD. Logarithm base 2 is
used as default. Use logbase = exp(1) for natural logarithm.}
}
\value{
The J divergence value for the given methylation levels is
returned
}
\description{
Given a the methylation levels from two individuals at a given
cytosine site, this function computes the J information divergence
(\emph{JD}) between methylation levels. The motivation to introduce
\emph{JD} in Methyl-IT is founded on:
\enumerate{
\item It is a symmetrised form of Kullback–Leibler divergence
(\eqn{D_{KL}}). Kullback and Leibler themselves actually defined
the divergence as: \eqn{D_{KL}(P || Q) + D_{KL}(Q || P)}, which
is symmetric and nonnegative, where the probability distributions
\emph{P} and \emph{Q} are defined on the same probability space
(see
\href{https://is.gd/oS8Bwv}{Wikipedia}).
\item In general, \emph{JD} is highly correlated with Hellinger
divergence, which is the main divergence currently used in
Methyl-IT (see examples for function
\code{\link{estimateDivergence}}.
\item By construction, the unit of measurement of \emph{JD} is:
bit of information, which set the basis for further
information-thermodynamics analyses.
}
}
\details{
The methylation levels \eqn{p_{ij}}, at a given cytosine site
\emph{i} from an individual \emph{j}, lead to the probability vectors
\eqn{p^{ij} = (p_{ij}, 1 - p_{ij})}. Then, the J-information divergence
between the methylation levels \eqn{p_{ij}} and the methylation levels
\eqn{q_i}, used as reference, corresponds to the divergence between the
probability vectors \eqn{p^{ij}} and \eqn{q^i = (q_i, 1 - q_i)}. \emph{JD}
is computed as in reference (1-2):

\deqn{JD(p^{ij}, q^i) = ((p_{ij} - q_i) * log(p_{ij}/q_i) +
                        (q_i - p_{ij}) * log((1 - p_{ij})/(1 - q_i)))/2}

Missing methylation levels, reported as NA or NaN, are replaced with zero.
}
\examples{
p <- c(0.5, 0.5)
estimateJDiv(p)

## A numerical trick is implemented. The J-divergence values are the
## same for the following vectors:
p <- c(0.9999999999999999, 0)
q <- c(1, 0)

estimateJDiv(p) == estimateJDiv(q)

}
\references{
\enumerate{
\item J. K. Chung, P. L. Kannappan, C. T. Ng, P. K. Sahoo, Measures of
distance between probability distributions. J. Math. Anal. Appl.
138, 280–292 (1989).
\item Lin J. Divergence Measures Based on the Shannon Entropy. IEEE
Trans Inform Theory, 1991, 37:145–51.
\item Sanchez R, Mackenzie SA. Information thermodynamics of cytosine
DNA methylation. PLoS One, 2016, 11:e0150427.
}
}
\seealso{
\url{https://en.wikipedia.org/wiki/Kullback-Leibler_divergence}
for more details, and \code{\link{estimateDivergence}} for an example of
using it.
}
\author{
Robersy Sanchez 11/27/2019 \url{https://github.com/genomaths}
}
