% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Extract.R
\name{Extract}
\alias{Extract}
\alias{[.InfDiv}
\alias{[.pDMP}
\alias{[.glmDataSet}
\alias{[.RangedGlmDataSet}
\title{Subscript operator for objects from several Methyl-IT classes}
\usage{
\method{[}{InfDiv}(x, i)

\method{[}{pDMP}(x, i)

\method{[}{glmDataSet}(x, i, j, ...)

\method{[}{RangedGlmDataSet}(x, i, j, ...)
}
\arguments{
\item{x}{A list-like or vector-like object}

\item{i}{Index specifying elements to Extract or replace. Index can be a
numeric or a character vector. See ?base::\code{\link[base]{Extract}} for
a description of these arguments.}
}
\value{
Same as in ?base::\code{\link[base]{Extract}}, but preserving
the class from the original object \eqn{x}.
}
\description{
\eqn{x[i]} returns an object of the length as \eqn{i} preserving
the class from the original object \eqn{x}. The following Methyl-IT
classes are considered: 'InfDiv','pDMP', 'glmDataSet', and
'RangedGlmDataSet'.
}
\seealso{
base::\code{\link[base]{Extract}}
}
\keyword{internal}
