% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitGGammaDist.R
\name{fitGGammaDist}
\alias{fitGGammaDist}
\title{Nonlinear fit of Generalized Gamma CDF (GGamma)}
\usage{
fitGGammaDist(x, parameter.values, location.par = FALSE,
  sample.size = 20, npoints = NULL, maxiter = 1024, ftol = 1e-12,
  ptol = 1e-12, maxfev = 1e+05, nlms = FALSE, verbose = TRUE, ...)
}
\arguments{
\item{x}{numerical vector}

\item{parameter.values}{initial parameter values for the nonlinear fit. If
the locator paramter is included (mu != 0), this must be given as
parameter.values = list(alpha = 'value', scale = 'value', mu = 'value',
psi = 'value') or if mu = 0, as: parameter.values =list(alpha = 'value',
scale = 'value', psi = 'value'). If not provided, then an initial guess
is provided.}

\item{location.par}{whether to consider the fitting to generalized gamma
distribution (GGamma) including the location parameter, i.e., a GGamma
with four parameters (GGamam4P).}

\item{sample.size}{size of the sample.}

\item{npoints}{number of points used in the fit. If the number of points if
greater than 10^6, then the fit is automatically set to npoints = 999999.
However, the reported values for R.Cross.val, AIC, and BIC are
computed taking into account the whole set of points.}

\item{maxiter}{positive integer. Termination occurs when the number of
iterations reaches maxiter. Default value: 1024.}

\item{ftol}{non-negative numeric. Termination occurs when both the actual
and predicted relative reductions in the sum of squares are at most ftol.
Therefore, ftol measures the relative error desired in the sum of
squares. Default value: 1e-12}

\item{ptol}{non-negative numeric. Termination occurs when the relative error
between two consecutive iterates is at most ptol. Therefore, ptol
measures the relative error desired in the approximate solution.
Default value: 1e-12.}

\item{maxfev}{integer; termination occurs when the number of calls to fn has
reached maxfev. Note that nls.lm sets the value of maxfev to
100*(length(par) + 1) if maxfev = integer(), where par is the list or
vector of parameters to be optimized.}

\item{nlms}{Logical. Whether to return the nonlinear model object
\code{\link[minpack.lm]{nls.lm}}. Default is FALSE.}

\item{verbose}{if TRUE, prints the function log to stdout}

\item{...}{arguments passed to or from other methods.}
}
\value{
Model table with coefficients and goodness-of-fit results:
    Adj.R.Square, deviance, AIC, R.Cross.val, and rho, as well as, the
    coefficient covariance matrix. If \strong{nlms = TRUE}, then a list with
    nonlinear model object \code{\link[minpack.lm]{nls.lm}} is returned.
}
\description{
This function performs the nonlinear fit of GGamma CDF of a
    variable x
}
\details{
The script algorithm tries to fit the three-parameter GGamma CDF
    ("GGamma3P") or the four-parameter GGamma ("GGamma4P") using a
    modification of Levenberg-Marquardt algorithm implemented in function
    'nls.lm' from 'minpack.lm' package that is used to perform the nonlinear
    fit. Cross-validations for the nonlinear regressions (R.Cross.val) were
    performed in each methylome as described in reference [1]. In addition,
    Stein's formula for adjusted R squared (rho) was used as an estimator of
    the average cross-validation predictive power [1].

    If the number of values to fit is >10^6, the fitting to a GGamma CDF
    would be a time consuming task. To reduce the computational time, the
    option summarized.data' can be set 'TRUE'. If npoint != NULL, the
    original variable values are summarized into 'npoint' bins and their
    midpoints are used as the new predictors. In this case, only the
    goodness-of-fit indicators AIC and R.Cross.val are estimated based on all
    the original variable x values.
}
\examples{
set.seed(123)
x <- rggamma(2000, alpha = 1.03, psi = 0.75, scale = 1.1)
fitGGammaDist(x)

}
\references{
1. Stevens JP. Applied Multivariate Statistics for the Social
    Sciences. Fifth Edit. Routledge Academic; 2009.
}
\author{
Robersy Sanchez - 06/03/2016
}
