% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getBaseMeansAndVariances.R
\name{.getBaseMeansAndVariances}
\alias{.getBaseMeansAndVariances}
\title{Calculate the means and variances by row}
\usage{
.getBaseMeansAndVariances(counts, sizeFactors)
}
\arguments{
\item{counts}{matrix of methylation counts}

\item{sizeFactors}{sizeFactors}
}
\value{
data.frame with means and variances by row
}
\description{
Calculate the means and variances in each row
}
\keyword{internal}
