% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getDMGs.R
\name{getDMGs}
\alias{getDMGs}
\alias{getDMGs,pDMP,GRanges-method}
\title{DMGs Estimation on Specified Genomic Region}
\usage{
getDMGs(dmps, regions, ...)

\S4method{getDMGs}{pDMP,GRanges}(
  dmps,
  regions,
  chr = NULL,
  colData,
  gene_id_col = NULL,
  gene_name_col = NULL,
  by.coord = FALSE,
  ignore.strand = TRUE,
  type = "within",
  only.hypo = FALSE,
  only.hyper = FALSE,
  num.cores = 1L,
  countFilter = TRUE,
  CountPerBp = NULL,
  minCountPerIndv = 3,
  maxGrpCV = NULL,
  FilterLog2FC = TRUE,
  pAdjustMethod = "BH",
  pvalCutOff = 0.05,
  MVrate = 0.98,
  Minlog2FC = 0.5,
  test = c("Wald", "LRT"),
  scaling = 1L,
  tasks = 0L,
  saveAll = FALSE,
  verbose = TRUE
)
}
\arguments{
\item{dmps}{An object from \emph{pDMP} or \emph{InfDiv} class, which are
generated by functions: \code{\link{getPotentialDIMP}} and
\code{\link{estimateDivergence}}, respectively.}

\item{regions}{A GRanges object with gene coordinates and gene IDs. A
column named \strong{'gene_id'} carrying the gene ids should be included in
the metacolumns. If the meta-column named 'gene_id' is not provided, then
gene (region) ids will be created using the gene (region) coordinates.}

\item{chr}{Optional. A character string denoting the chromosome from
the \emph{regions} (i.e., the \code{\link[GenomeInfoDb]{seqnames}}) where
the DMGs will be estimated.}

\item{colData}{The same as in \code{\link{glmDataSet}}. A data frame with
one column named 'condition', which must be a factor with exactly two
levels. The rownames of \emph{colData} individual samples. The row names of
\emph{colData} must correspond to the column names of the count matrix.}

\item{gene_id_col}{Optional. An integer denoting the column from the
\emph{GENES} metacolumn where the gene ids are given.}

\item{gene_name_col}{Optional. An integer denoting the column from the
\emph{GENES} metacolumn where the gene 'name' are given.}

\item{by.coord}{logical(1). If TRUE, then the DMP are count per coordinate
and not per gene id.}

\item{ignore.strand, type}{Same as for
\code{\link[GenomicRanges]{findOverlaps-methods}}.}

\item{only.hypo, only.hyper}{logical(1). Whether to select only
hypo-methylated or hyper-methylated cytosine sites.}

\item{num.cores, countFilter, CountPerBp, minCountPerIndv, maxGrpCV}{The same
as in function \code{\link{countTest2}}.}

\item{FilterLog2FC, pAdjustMethod, pvalCutOff, MVrate, Minlog2FC}{The same
as in function \code{\link{countTest2}}.}

\item{test, scaling, tasks, saveAll, verbose}{The same
as in function \code{\link{countTest2}}.}
}
\description{
This function is a wrapping of several functions called in
through the steps to estimate differentially methylated regions (DMGs).
}
\seealso{
\code{\link{countTest2}}, \code{\link{getDIMPatGenes}},
and \code{\link{glmDataSet}}.
}
