% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getMethContext.R
\name{getMethContext}
\alias{getMethContext}
\title{Get Methylation Context from a chromosome DNA sequence}
\usage{
getMethContext(chr.seq, chromosome, verbose = TRUE)
}
\arguments{
\item{chr.seq}{DNA sequence from a chromosome in fasta format.}

\item{chromosome}{Chromosome name.}

\item{verbose}{If TRUE, prints the function log to stdout}
}
\value{
GRanges object with three columns: 'trinucleotide', methylation
    context, and 'CHH' methylation subcontexts: 'CHA', 'CHC', and 'CHT'.
}
\description{
This function retrieves the methylation context from a
    chromosome DNA sequence in fasta format.
}
\examples{
dna <- Biostrings::DNAString(x = "CCCTAACGACCCTAACGCTACCCTAAACCTCTGAAT",
    start = 1, nchar = NA)
getMethContext(chr.seq = dna, chromosome = "1", verbose = TRUE)


}
