% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggamma.R
\name{ggamma}
\alias{ggamma}
\alias{dggamma}
\alias{rggamma}
\alias{pggamma}
\alias{qggamma}
\title{Generalized Gamma distribution}
\usage{
dggamma(q, alpha = 1, scale = 1, mu = 0, psi = 1, log.p = FALSE)

pggamma(
  q,
  alpha = 1,
  scale = 1,
  mu = 0,
  psi = 1,
  lower.tail = TRUE,
  log.p = FALSE
)

qggamma(
  p,
  alpha = 1,
  scale = 1,
  mu = 0,
  psi = 1,
  lower.tail = TRUE,
  log.p = FALSE
)

rggamma(n, alpha = 1, scale = 1, mu = 0, psi = 1)
}
\arguments{
\item{q}{numeric vector.}

\item{alpha}{numerical parameter, strictly positive (default 1). The
generalized gamma becomes the gamma distribution for alpha = 1.}

\item{scale, psi}{the same real positive parameters as is used for the Gamma
distribution. These are numerical and strictly positives; default 1.
(see ?pgamma).}

\item{mu}{location parameter (numerical, default 0).}

\item{log.p}{logical; if TRUE, probabilities/densities p are returned as
log(p).}

\item{lower.tail}{logical; if TRUE (default), probabilities are
\eqn{P(X<=x)}, otherwise, \eqn{P(X > x)}}

\item{p}{vector of probabilities.}

\item{n}{number of observations.}
}
\value{
GG PDF values (3-parameters or 4-parameters) for dggamma,
GG probability for pggamma, quantiles or GG random generated values for
rggamma.
}
\description{
Probability density function (PDF), cumulative density function
(CDF), quantile function and random generation for the Generalized Gamma
(GG) distribution with 3 or 4 parameters: alpha, scale, mu, and psi. The
function is reduced to GGamma distribution with 3 parameters by setting
mu = 0.
}
\details{
Details about these function can be found in references 1 to 3. You
may also see section Note at ?pgamma or ?rgamma. Herein, we are using Stacy'
s formula (references 2 to 3) with the parametrization given in reference 4
(equation 6, page 12). As in the case of gamma distribution function, the
cumulative distribution function (as given in equation 12, page 13 from
reference 4) is expressed in terms of the lower incomplete gamma function
(see ?pgamma).

The GG distribution with parameters \eqn{\alpha}, \eqn{\beta} (scale),
\eqn{\psi}, and \eqn{\mu} has density:

\deqn{f(x | \alpha, \beta, \mu, \psi) = \alpha exp(-((x-\mu)/
    \beta)^\alpha) ((x-\mu)/\beta)^(\alpha * \psi - 1)/(\beta Gamma(\psi))}
}
\examples{
q <- (1:9)/10
pggamma(q, alpha = 1, scale = 1, mu = 0,
        psi = 1, lower.tail = TRUE, log.p = FALSE)

## To fit random generated numbers
set.seed(123)
x <- rggamma(2000, alpha = 1.03, psi = 0.75, scale = 2.1)
fitGGammaDist(x)

}
\references{
\enumerate{
\item Handbook on  STATISTICAL DISTRIBUTIONS for experimentalists
(p. 73) by Christian Walck. Particle Physics Group Fysikum. University of
Stockholm (e-mail: walck@physto.se )
\item Stacy, E. W. A Generalization of the Gamma Distribution. Ann. Math. Stat.
33, 1187–1192 (1962).
\item Stacy E, Mihram G (1965) Parameter estimation for a generalized gamma
distribution. Technometrics 7: 349-358.
\item Sanchez, R. & Mackenzie, S. A. Information Thermodynamics of Cytosine DNA
Methylation. PLoS One 11, e0150427 (2016).
}
}
