% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glmDataSet.R
\name{glmDataSet}
\alias{glmDataSet}
\title{Data set constructor for class glmDataSet}
\usage{
glmDataSet(GR = NULL, counts = NULL, colData = NULL)
}
\arguments{
\item{GR}{A GRanges object with the count matrix of DMPs in the metacolumns
(see \emph{'counts'}). If provided, then leave paramater
\emph{'counts = NULL'}.}

\item{counts}{Count matrix of DMPs with minimal dimmensions 1 (row) x 4
(columns). Column names must corresponds to the rownames from parameter
 'colData'.}

\item{colData}{A data frame with one columnn named 'condition', which must be
a factor with exactly two levels. The rownames of \emph{colData}
individual samples. The row names of \emph{colData} must correspond to
th column names of the count matrix.}
}
\value{
A RangedGlmDataSet object, good for downstream use with
    Methyl-IT \code{link{countTest2}} function.
}
\description{
This function is used to build a object suitable to be used
    with Methyl-IT \code{link{countTest2}} function.
}
\details{
Data set constructor for class glmDataSet also validate the object
}
\examples{
DMPs <- data.frame(chr = "chr1", start = 1:10,
                   end = 1:10,strand = '*',
                   treat1 = rnbinom(size = 10, mu = 4, n = 500),
                   treat2 = rnbinom(size = 10, mu = 4, n = 500),
                   cntrl1 = rnbinom(size = 10, mu = 4, n = 500),
                   cntrl2 = rnbinom(size = 10, mu = 4, n = 500))
DMPs <- makeGRangesFromDataFrame(DMPs, keep.extra.columns = TRUE)
condition <- data.frame(condition = factor(c("TT","TT","CT","CT"),
                        levels = c("CT", "TT")))
rownames(condition) <- names(mcols(DMPs))
DIMR <- glmDataSet(GR = DMPs, colData = condition)

}
\author{
Robersy Sanchez
}
