% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gmMean.R
\name{.gmMean}
\alias{.gmMean}
\title{Calculate the geometric mean}
\usage{
.gmMean(x, na.rm = TRUE)
}
\arguments{
\item{x}{numeric vector}

\item{na.rm}{Logical value. If TRUE, the NA values will be removed}
}
\value{
geometric mean
}
\description{
Calculate the geometric mean
}
\keyword{internal}
