% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_gof.R
\docType{data}
\name{gof}
\alias{gof}
\title{Simulated dataset of nonlinear fits used in the examples}
\format{
gof is a list of best fitted nonlinear probability distribution
model estimated for the dataset 'HD' (available in MethylIT package)
'gof' carries the information on the best fitted probability
distribution model for each individual sample. 'gof' was obtained with
function \code{\link{gofReport}}.
}
\usage{
gof
}
\description{
Each individuals sample includes 10000 cytosine positions
}
\keyword{datasets}
