% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gunzip.R
\name{.gunzip}
\alias{.gunzip}
\title{Decompress 'gzip' files}
\usage{
.gunzip(
  filename,
  destname = gsub("[.]gz$", "", filename),
  overwrite = FALSE,
  remove = TRUE,
  BFR.SIZE = 1e+07
)
}
\arguments{
\item{filename}{Character vector with the files names to be decompressed}

\item{destname}{The destination file.}

\item{overwrite}{Logic (FALSE) indicating whether or not to overwrite a
destfile of the same name.}

\item{remove}{Logic (TRUE) indicating whether or not to remove the original
file after completion.}

\item{BFR.SIZE}{The size of the read buffer.}
}
\value{
Invisibly, the number of bytes read.
}
\description{
Decompress 'gzip' files.
}
\details{
This function is originally from the R packages 'R.utils'. Herein,
the version as given by 'GEOquery' is incorporated.
}
\author{
Original author: Henrik Bengtsson.
}
\keyword{internal}
