% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_hdiv.R
\docType{data}
\name{hdiv}
\alias{hdiv}
\title{Simulated dataset of Information divergences used in the examples}
\format{
'hdiv is an object from class 'InfDiv' with
\code{\link[GenomicRanges]{GRanges-class}} ascelements carrying in the
meta-columns the following variables:
\describe{
\item{c1}{Methylated read counts from the reference sample.}
\item{t1}{Unmethylated read counts from the reference sample.}
\item{c2}{Methylated read counts from the treatment sample.}
\item{t2}{Unmethylated read counts from the treatment sample.}
\item{p1}{methylation level from the reference sample.}
\item{p2}{methylation level from the treatment sample.}
\item{TV}{the total variation distance (difference of
methylation levels)}
\item{bay.TV}{TV computes with Bayesian correction for the
methelation levels.}
\item{hdiv}{Hellinger divergence.}
\item{jdiv}{J divergence.}
}
'hdiv' was obtained with function \code{\link{estimateDivergence}}.
}
\usage{
hdiv
}
\description{
Each individuals sample includes 10000 cytosine positions
}
\keyword{datasets}
