% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helmholtz_free_energy.R
\name{helmholtz_free_energy}
\alias{helmholtz_free_energy}
\alias{helmholtz_free_energy,missingORNULL-method}
\alias{helmholtz_free_energy,cdfMODEL-method}
\alias{helmholtz_free_energy,cdfMODELlist-method}
\alias{helmholtz_free_energy,ProbDistrList-method}
\title{Helmholtz Free Energy of Generalized Gamma Distribution}
\usage{
helmholtz_free_energy(model, R = 8.31446261815324, T.abs = 310.15, ...)

\S4method{helmholtz_free_energy}{missingORNULL}(
  model,
  R = 8.31446261815324,
  T.abs = 310.15,
  pars,
  log.base = exp(1),
  kj = FALSE,
  terms = FALSE
)

\S4method{helmholtz_free_energy}{cdfMODEL}(
  model,
  R = 8.31446261815324,
  T.abs = 310.15,
  log.base = exp(1),
  kj = FALSE,
  terms = FALSE
)

\S4method{helmholtz_free_energy}{cdfMODELlist}(
  model,
  R = 8.31446261815324,
  T.abs = 310.15,
  log.base = exp(1),
  kj = FALSE,
  terms = FALSE
)

\S4method{helmholtz_free_energy}{ProbDistrList}(
  model,
  R = 8.31446261815324,
  T.abs = 310.15,
  log.base = exp(1),
  kj = FALSE,
  terms = FALSE
)
}
\arguments{
\item{model}{An object from any of the classes created in MethylIT
pipeline: \strong{\emph{cdfMODEL}}, \emph{\strong{cdfMODELlist}}, or
\emph{\strong{ProbDistrList}}. If given, then the parameter values are taken
from the model.}

\item{R}{A number or NULL. The gas constant is given as default values,
which is proportionality constant that relates the energy scale in physics
to the temperature scale and the scale used for amount of substance.}

\item{T.abs}{Absolute temperature (in Kelvin units). Default is
T.abs = 310.15.}

\item{pars}{Optional. A numerical vector containing the model parameter
values in the given in order: \strong{\emph{alpha, scale}}, and
\strong{\emph{delta}}.}

\item{log.base}{A positive number. The same as in function
\code{\link[base]{log}}.}

\item{kj}{Logical. Whether to return the result in KJ * mol^-1.}

\item{terms}{logical(1). If term = TRUE, then a numerical vector with terms
contributing to the Helmholtz free energy are provided.}
}
\value{
Helmholtz free energy of system according to the model. If R is the
gas constant and T.abs given in Kelvin units, then the energy is returned in
Joule per mol units.
}
\description{
This function computes the Helmholtz Free Energy (HFE) for
member of the Generalized Gamma (GG) Distribution family. The Helmholtz free
energy  represents the driving force for NVT systems, the thermodynamic
potential that measures the “useful” work obtainable from a closed
thermodynamic system at a constant temperature and volume.

GG density is given as:

\eqn{exp(-y^\alpha)*\alpha*y^(\alpha*\delta - 1)/(scale*\gamma(\delta))}

(see \href{https://goo.gl/1n4kpW}{(Wikipedia)})

A list of the member of the GG distribution family with the corresponding
parameter settings is provided in Table 1 from reference (1). For example, to
compute the HFE of a Weibull distribution model, we just set: \eqn{alpha > 0}
and \eqn{delta = 1} (notice that parameter are named different in reference
(1)).
}
\details{
The values of the constants R and T.abs can be simply 1, which
implies dimensionless units units, with perhaps no physical interpretation,
only carrying informational meaning, i.e., only meaningful in the framework
of information theory.
}
\examples{
## Loading the probability distribution models
data(gof, "MethylIT")
## Helmholtz free energy at 310.15 K (Kelvin) in J * mol^-1
helmholtz_free_energy (gof)
}
\references{
\enumerate{
\item Crooks, Gavin E. (2015) The Amoroso Distribution.
\href{https://arxiv.org/pdf/1005.3274.pdf}{arXiv:1005.3274v2}.
}
}
\seealso{
\code{\link{gibb_entropy}}
}
