% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_lda_perf.R
\docType{data}
\name{lda_perf}
\alias{lda_perf}
\title{Classification LDA model for simulated dataset of DMPs used in examples}
\format{
\emph{lda_perf} is list object consisting of the following elements:
\describe{
\item{\strong{Performance}}{Classification performance of the
\emph{"LDA"} model on the set of \code{\link{dmps}}}.
\item{\strong{FDR}}{False discovery rate estimated for the \emph{"LDA"}
model on the set of \code{\link{dmps}}}
\item{\strong{model}}{The \code{\link[MASS]{lda}} model fitted on the set
of \code{\link{dmps}}}.
}

\strong{\emph{lda_perf}} is a list object obtained applying
\code{\link{evaluateDIMPclass}} function on the set of
\code{\link{dmps}}.
}
\usage{
lda_perf
}
\description{
This data/object carries the information about the classification performance
of a Linear Discriminant (\emph{"LDA"})) model on the set of
\code{\link{dmps}}.
}
\keyword{datasets}
