% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_logit_perf.R
\docType{data}
\name{logit_perf}
\alias{logit_perf}
\title{Classification logistic model for simulated dataset of DMPs used in examples}
\format{
\emph{logit_perf} is an object from class \emph{"LogisticR"},
consisting of a list with the following elements:
\describe{
\item{\strong{Performance}}{Classification performance of the
\emph{"logistic"} model on the set of \code{\link{dmps}}}.
\item{\strong{FDR}}{False discovery rate estimated for the logistic
model on the set of \code{\link{dmps}}}
\item{\strong{model}}{The fitted \code{\link[stats]{glm}} logistic
model fitted on the set of \code{\link{dmps}}}.
}

\strong{\emph{logit_perf}} is an object from "LogisticR" class, which was
obtained applying \code{\link{evaluateDIMPclass}} function on the set of
\code{\link{dmps}}.
}
\usage{
logit_perf
}
\description{
This data/object carries the information about the classification performance
of \emph{"logistic"} model on the set of \code{\link{dmps}}.
}
\keyword{datasets}
