% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/machine_ent.R
\name{machine_ent}
\alias{machine_ent}
\alias{machine_ent,missing_OR_NULL-method}
\alias{machine_ent,cdfMODEL-method}
\alias{machine_ent,cdfMODELlist-method}
\alias{machine_ent,ProbDistrList-method}
\title{Molecular Machine Entropy}
\usage{
machine_ent(model, R = 8.31446261815324, ...)

\S4method{machine_ent}{missing_OR_NULL}(model, R = 8.31446261815324, pars, only.mchent = FALSE)

\S4method{machine_ent}{cdfMODEL}(model, R = 8.31446261815324, only.mchent = FALSE)

\S4method{machine_ent}{cdfMODELlist}(model, R = 8.31446261815324, only.mchent = FALSE)

\S4method{machine_ent}{ProbDistrList}(model, R = 8.31446261815324, only.mchent = FALSE)
}
\arguments{
\item{model}{An object from any of the classes created in MethylIT
pipeline: \emph{\strong{cdfMODEL}}, \emph{\strong{cdfMODELlist}}, or
\emph{\strong{ProbDistrList}}. If given, then the parameter values are taken
from the model.}

\item{R}{A number or NULL. The gas constant
(\eqn{R = 8.31446261815324 J * (K * mol)^-1}) is given as default value,
which is proportionality constant that relates the energy scale in physics
to the temperature scale and the scale used for amount of substance.}

\item{pars}{Optional. A numerical vector containing the model parameter
values in the given in order: \strong{\emph{alpha, scale}}, and
\strong{\emph{delta}}.}
}
\value{
Molecular machine contribution to the methylation entropy based on
the model provided.
}
\description{
This function computes the contribution of a molecular
machine (methylation molecular machinery) to the entropy of individual
methylation system, which is estimated for members of the
Generalized Gamma (GG) Distribution family. GG density is given as:

\eqn{exp(-y^\alpha) * \alpha*y^{\alpha*\psi - 1}/(scale*\gamma(\psi))}

(see \href{https://goo.gl/1n4kpW}{(Wikipedia)})

A list of the member of the GG distribution family with the corresponding
parameter settings is provided in Table 1 from reference (1). For example, to
compute the Gibbs entropy of a Weibull distribution model, we just set:
\eqn{alpha > 0} and \eqn{delta = 1} (notice that parameter are named
different in reference (1)).
}
\details{
The entropy contribution derives from the GG model as given by
functions \code{\link{nonlinearFitDist}} and \code{\link{gofReport}}. Given
the model parameters \eqn{\alpha} and \eqn{\psi} the molecular machine
contribution to the methylation system's entropy is given by the equation:

\deqn{digamma(\psi) * ((1/\alpha) - \psi) + \psi}

where \code{\link[base]{digamma}} is the first derivative of the logarithm of
the gamma function
(\href{https://en.wikipedia.org/wiki/Digamma_function}{Wikipedia}).

The value of the constant R can be simply 1, which returns the
Shannon entropy in bit units, only carrying informational meaning.
}
\references{
\enumerate{
\item Crooks, Gavin E. (2015) The Amoroso Distribution.
\href{https://arxiv.org/pdf/1005.3274.pdf}{arXiv:1005.3274v2}.
}
}
