% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meth_levels.R
\name{meth_level}
\alias{meth_level}
\alias{meth_level,GRanges-method}
\alias{meth_level,list-method}
\title{Compute methylation levels}
\usage{
meth_level(
  GR,
  Bayesian = FALSE,
  columns = c(mC1 = 1, uC1 = 2, mC2 = 3, uC2 = 4),
  min.coverage = 4,
  tv = FALSE,
  bay.tv = FALSE,
  preserve.gr = FALSE,
  num.cores = 1,
  tasks = 0L,
  verbose = TRUE,
  ...
)

\S4method{meth_level}{GRanges}(
  GR,
  Bayesian = FALSE,
  columns = c(mC1 = 1, uC1 = 2, mC2 = 3, uC2 = 4),
  min.coverage = 4,
  tv = FALSE,
  bay.tv = FALSE,
  preserve.gr = FALSE,
  num.cores = 1,
  tasks = 0L,
  verbose = TRUE,
  ...
)

\S4method{meth_level}{list}(
  GR,
  Bayesian = FALSE,
  columns = c(mC1 = 1, uC1 = 2, mC2 = 3, uC2 = 4),
  min.coverage = 4,
  tv = FALSE,
  bay.tv = FALSE,
  preserve.gr = FALSE,
  num.cores = 1,
  tasks = 0L,
  verbose = TRUE,
  ...
)
}
\arguments{
\item{GR}{A matrix of counts or \code{\link{GenomicRanges}{"GRanges"}}
object with the table of counts in the meta-columns (methylated mC and
unmethylated uC cytosines) or list of matrices of GRanges objects . Unless
specified in the parameter 'columns', the methylation counts must be given in
the first four columns: 'mC1' and 'uC1' methylated and unmethylated counts
for control sample, and 'mC2' and 'uC2' methylated and unmethylated counts
for treatment sample, respectively.}

\item{Bayesian}{logical(1). Whether to perform the estimations based on
posterior estimations of methylation levels.}

\item{min.coverage}{An integer or an integer vector of length 2. Cytosine
sites where the coverage in both samples, 'x' and 'y', are less than
min.coverage' are discarded. The cytosine site is preserved, however, if the
coverage is greater than 'min.coverage' in at least one sample. If
'min.coverage' is an integer vector, then the corresponding min coverage is
applied to each sample.}

\item{tv}{logical(1). Whether to compute the total variation distance at each
cytosine site. That is, the difference of methylation levels.}

\item{bay.tv}{logical(1). Whether to compute the total variation distance at
each cytosine site based on Bayesian estimation of methylation levels.}

\item{preserve.gr}{logical(1). Option of whether to preserve all
the metadata from the original GRanges object.}

\item{num.cores, tasks}{Parameters for parallel computation using package
\code{\link[BiocParallel]{BiocParallel-package}}: the number of cores to use,
i.e. at most how many child processes will be run simultaneously (see
\code{\link[BiocParallel]{bplapply}} and the number of tasks per job (only
for Linux OS). These parameters will be passed to
\code{\link[GenomicWordFramework]{uniqueGRanges}}.}

\item{verbose}{if TRUE, prints the function log to stdout}
}
\description{
This function computes the
}
