% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meth_levels.R
\name{meth_levels}
\alias{meth_levels}
\alias{meth_levels,ANY,data.frame-method}
\alias{meth_levels,GRanges,ANY-method}
\alias{meth_levels,list,ANY-method}
\title{Compute methylation levels}
\usage{
meth_levels(
  GR,
  x,
  columns = c(mC1 = 1, uC1 = 2, mC2 = NULL, uC2 = NULL),
  Bayesian = FALSE,
  min.coverage = 4,
  tv = FALSE,
  bay.tv = FALSE,
  filter = FALSE,
  preserve.dt = FALSE,
  num.cores = 1,
  tasks = 0L,
  verbose = TRUE,
  ...
)

\S4method{meth_levels}{ANY,data.frame}(
  GR,
  x,
  columns = c(mC1 = 1, uC1 = 2, mC2 = NULL, uC2 = NULL),
  Bayesian = FALSE,
  min.coverage = 4,
  tv = FALSE,
  bay.tv = FALSE,
  filter = FALSE,
  preserve.dt = FALSE,
  num.cores = 1,
  tasks = 0L,
  verbose = TRUE,
  ...
)

\S4method{meth_levels}{GRanges,ANY}(
  GR,
  x,
  columns = c(mC1 = 1, uC1 = 2, mC2 = NULL, uC2 = NULL),
  Bayesian = FALSE,
  min.coverage = 4,
  tv = FALSE,
  bay.tv = FALSE,
  filter = FALSE,
  preserve.dt = FALSE,
  num.cores = 1,
  tasks = 0L,
  verbose = TRUE,
  ...
)

\S4method{meth_levels}{list,ANY}(
  GR,
  x = NULL,
  columns = c(mC1 = 1, uC1 = 2, mC2 = 0, uC2 = 0),
  Bayesian = FALSE,
  min.coverage = 4,
  tv = FALSE,
  bay.tv = FALSE,
  filter = FALSE,
  preserve.dt = FALSE,
  num.cores = 1,
  tasks = 0L,
  verbose = TRUE,
  ...
)
}
\arguments{
\item{GR, x}{A \code{\link[GenomicRanges]{GRanges-class}} object
(\strong{\emph{GR}}) or 'data.frame' (\strong{\emph{x}}) with a matrix of
counts in the meta-columns (methylated mC and unmethylated uC cytosines) or a
list of \code{\link[GenomicRanges]{GRanges-class}} objects.}

\item{columns}{Vector of one or two integer numbers denoting the indexes of
the columns where the methylated and unmethylated read counts are found.
Unless specified in the parameter 'columns', the methylation counts must be
given in the first four columns: 'mC1' and 'uC1' methylated and unmethylated
counts for control sample, and 'mC2' and 'uC2' methylated and unmethylated
counts for treatment sample, respectively.}

\item{Bayesian}{logical(1). Whether to perform the estimations based on
posterior estimations of methylation levels.}

\item{min.coverage}{An integer or an integer vector of length 2. Cytosine
sites where the coverage in both samples, 'x' and 'y', are less than
min.coverage' are discarded. The cytosine site is preserved, however, if the
coverage is greater than 'min.coverage' in at least one sample. If
'min.coverage' is an integer vector, then the corresponding min coverage is
applied to each sample.}

\item{tv}{logical(1). Whether to compute the total variation distance at each
cytosine site. That is, the difference of methylation levels.}

\item{bay.tv}{logical(1). Whether to compute the total variation distance at
each cytosine site based on Bayesian estimation of methylation levels.}

\item{preserve.dt}{logical(1). Option of whether to preserve all
the metadata from the original 'data.frame' or
\code{\link[GenomicRanges]{GRanges-class}} object.}

\item{num.cores, tasks}{Parameters for parallel computation using package
\code{\link[BiocParallel]{BiocParallel-package}}: the number of cores to use,
i.e. at most how many child processes will be run simultaneously (see
\code{\link[BiocParallel]{bplapply}} and the number of tasks per job (only
for Linux OS). These parameters will be passed to
\code{\link{uniqueGRanges}}.}

\item{verbose}{if TRUE, prints the function log to stdout}
}
\description{
This function computes the
}
