% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MethylIT.R
\name{MethylIT}
\alias{MethylIT}
\title{MethylIT: Methylation Analysis Based on Signal Detection and Machine
Learning}
\description{
Methyl-IT implements methylation analysis based on signal
detection theory and machine-learning. Methylation changes are expressed in
terms of information divergence (ID) of methylation level. Methylation
process is a stochastic process, particularly, a biochemical-biophysical
process which must not be reduced to statistic. Methyl-IT includes the
information on the statistical biophysics of the methylation process with
the estimation of the probability distribution of the methylation
background noise (plus signal), which is used as null distribution for the
application of basic signal detection theory.
}
