% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlCutpoint.R
\name{mlCutpoint}
\alias{mlCutpoint}
\title{Machine-learning Approach for Cutpoint Estimation}
\usage{
mlCutpoint(
  LR,
  control.names,
  treatment.names,
  column,
  div.col,
  tv.col = NULL,
  tv.cut,
  post.cut = 0.5,
  classifier1,
  classifier2 = NULL,
  n.pc,
  prop = 0.6,
  stat = 0L,
  cut.values = NULL,
  num.cores = 1L,
  tasks = 0L,
  ...
)
}
\arguments{
\item{LR, res, control.names, treatment.names, column, div.col}{Same as in
\code{\link{estimateCutPoint}}}

\item{column, div.col, tv.col, tv.cut, cut.values, stat}{Same as in
\code{\link{estimateCutPoint}}}

\item{classifier1, classifier2, n.pc, prop, post.cut}{Same as in
\code{\link{estimateCutPoint}}}

\item{cutp_data, num.cores, tasks}{Same as in \code{\link{estimateCutPoint}}}
}
\value{
Specified in function \code{\link{estimateCutPoint}} for parameter
setting \emph{simple = FALSE}
}
\description{
Internal function to estimate cutpoint following a
machine-learning approach
}
\details{
This function is called by function \code{\link{estimateCutPoint}}.
}
\examples{
## Get a set of potential DMPS (PS)
data(PS, package = 'MethylIT')

cutp <- mlCutpoint(LR = PS,
                 column = c(hdiv = TRUE, TV = TRUE,
                            wprob = TRUE, pos = TRUE),
                 classifier1 = 'qda', n.pc = 4,
                 control.names = c('C1', 'C2', 'C3'),
                 treatment.names = c('T1', 'T2', 'T3'),
                 tv.cut = 0.68, prop = 0.6,
                 div.col = 9L)
cutp
}
\keyword{internal}
