% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_nlms.R
\docType{data}
\name{nlms}
\alias{nlms}
\title{Simulated dataset of nonlinear fits used in the examples}
\format{nlms is a list of best fitted nonlinear probability distribution
    model estimated for the datset 'HD' (available in MethylIT pacakga)
    'nlms' carries the information on the best fitted probability
    distribution model for each individual sample. 'nlms' was obtianed with
    function \code{\link{nonlinearFitDist}}.}
\usage{
nlms
}
\description{
Each individuals sample includes 5000 cytosine positions
}
\keyword{datasets}
