% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pcaQDA.R
\name{pcaQDA}
\alias{pcaQDA}
\alias{predict.pcaQDA}
\title{Quadratic Discriminant Analysis (QDA) using Principal Component
Analysis (PCA)}
\usage{
pcaQDA(
    formula = NULL,
    data = NULL,
    grouping = NULL,
    n.pc = 1,
    scale = FALSE,
    center = FALSE,
    tol = 1e-04,
    method = "moment",
    max.pc = NULL
)

\method{predict}{pcaQDA}(
    object,
    newdata,
    type = c("qda.pred", "class", "posterior", "pca.ind.coord", "all"),
    ...
)
}
\arguments{
\item{formula}{Same as in \code{\link[MASS]{qda}} from package 'MASS'.}

\item{data}{Same as in \code{\link[MASS]{qda}} from package 'MASS'.}

\item{grouping}{Same as in \code{\link[MASS]{qda}} from package 'MASS'.}

\item{n.pc}{Number of principal components to use in the qda.}

\item{scale}{Same as in \code{\link[stats]{prcomp}} from package 'stats'.}

\item{center}{Same as in \code{\link[stats]{prcomp}} from package 'stats'.}

\item{tol}{Same as in \code{\link[stats]{prcomp}} from package 'stats'.}

\item{method}{Same as in \code{\link[MASS]{qda}} from package 'MASS'.}

\item{max.pc}{Same as in parameter 'rank.' from
\code{\link[stats]{prcomp}} from package 'stats'.}

\item{object}{To use with function 'predict'. A 'pcaQDA' object containing a
list of two objects: 1) an object of class inheriting from 'qda' and 2)
an object of class inheriting from 'prcomp'.}

\item{newdata}{To use with function 'predict'. New data for classification
prediction.}

\item{type}{To use with function 'predict'. The type of prediction
required. The default is 'all' basic predictions: classes and posterior
classification probabilities. Option 'qda.pred' returns the
object given by function 'predict.qda' from MASS package: 'class',
'posterior', 'scores' (cases scores on discriminant variables,
see \code{\link[MASS]{qda}}.}

\item{...}{Not in use.}
}
\value{
Function 'pcaQDA' returns an object ('pcaQDA') consisting of a list with
two objects:
\enumerate{
\item 'qda': an object of class \code{\link[MASS]{qda}} from package
'MASS'.
\item 'pca': an object of class \code{\link[stats]{prcomp}} from package
'stats'.
}
For information on how to use these objects see ?qda and ?prcomp.
}
\description{
The principal components (PCs) for predictor variables provided
as input data are estimated and then the individual coordinates in the
selected PCs are used as predictors in the qda

Predict using a PCA-LDA model built with function 'pcaLDA'
}
\details{
The principal components (PCs) are obtained using the function
'prcomp' from R package 'stats', while the qda is performed using the
'qda' function from R package 'MASS'. The current application only uses
basic functionalities of mentioned functions. As shown in the example,
'pcaQDA' function can be used in general classification problems.
}
\examples{
data(iris)
qd1 <- pcaQDA(formula = Species ~ Petal.Length + Sepal.Length + Sepal.Width,
data = iris, n.pc = 1, max.pc = 2, scale = TRUE, center = TRUE)
## === Prediction === ##
qd2 <- pcaQDA(formula = Species ~., data = iris, n.pc = 1, max.pc = 2,
scale = TRUE, center = TRUE)

## Set a random seed to reproduce the sampling
set.seed(123)
idx <- sample.int(150, 40)
newdata <- iris[idx, 1:4]
newdata.prediction <- predict(qd2, newdata = newdata, type = 'all')

## The confusion matrix
x <- data.frame(TRUE.class = iris$Species[idx],
PRED.class = newdata.prediction$class)
table(x)
}
\seealso{
\code{\link{pcaLDA}}, \code{\link[MASS]{qda}} and
\code{\link[MASS]{predict.lda}}
}
