% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.ldaDMP.R
\name{predict.ldaDMP}
\alias{predict.ldaDMP}
\title{Predict function for 'ldaDMP' method}
\usage{
\method{predict}{ldaDMP}(
  object,
  newdata,
  type = c("class", "posterior", "scores", "lda.pred"),
  ...
)
}
\arguments{
\item{object}{To use with function 'predict'. An object from class 'ldaDMP'}

\item{newdata}{To use with function 'predict'. New data for classification
prediction}

\item{type}{To use with function 'predict'. The type of prediction
required.  The default is "all" basic predictions: classes and posterior
classification probabilities. Option "lda.pred" returns the
object given by function 'predict.lda' from MASS package: 'class',
'posterior', 'scores' (cases scores on discriminant variables,
see \code{\link[MASS]{lda}}.}

\item{...}{arguments passed to or from other methods.}
}
\description{
Predict using a LDA model built with function
\code{\link[MethylIT]{evaluateDIMPclass}}
}
\seealso{
\code{\link[MethylIT]{estimateCutPoint}}, \code{\link[MASS]{lda}}
}
\keyword{internal}
