% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.qdaDMP.R
\name{predict.qdaDMP}
\alias{predict.qdaDMP}
\alias{predict.pcaqda}
\title{Predict function for 'qdaDMP' method}
\usage{
predict.qdaDMP(
  object,
  newdata,
  type = c("all", "class", "posterior", "scores", "qda.pred"),
  ...
)
}
\arguments{
\item{object}{To use with function 'predict'. An object from class 'qdaDMP'}

\item{newdata}{To use with function 'predict'. New data for classification
prediction}

\item{type}{To use with function 'predict'. The type of prediction
required. The default is 'all' basic predictions: classes and posterior
classification probabilities. Option 'qda.pred' returns the
object given by function 'predict.qda' from MASS package: 'class',
'posterior', 'scores' (cases scores on discriminant variables,
see \code{\link[MASS]{qda}}.}

\item{...}{arguments passed to or from other methods.}
}
\description{
Predict using a qda model built with function
\code{\link[MethylIT]{evaluateDIMPclass}}
}
\seealso{
\code{\link[MethylIT]{estimateCutPoint}}, \code{\link[MASS]{qda}}
}
\keyword{internal}
