% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.CutPoint.R
\name{print.CutPoint}
\alias{print.CutPoint}
\title{Printing object from 'CutPoint' class by simple print methods}
\usage{
\method{print}{CutPoint}(x, digits = getOption("digits"))
}
\arguments{
\item{x}{Object of class \code{"CutPoint"}}

\item{digits}{Number of significant digits to be used.}
}
\description{
Printing object from 'CutPoint' class by simple print methods
}
\keyword{internal}
