% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.glmDataSet.R
\name{print.glmDataSet}
\alias{print.glmDataSet}
\title{Printing object from \emph{glmDataSet} and \emph{"RangedGlmDataSet"}
classes by simple print methods.}
\usage{
\method{print}{glmDataSet}(x, digits = getOption("digits"), ...)
}
\arguments{
\item{x}{Object from class \emph{glmDataSet} or from class
\emph{"RangedGlmDataSet"}.}

\item{digits}{Number of significant digits to be used.}
}
\description{
Printing object from \emph{glmDataSet} and \emph{"RangedGlmDataSet"}
classes by simple print methods.
}
\keyword{internal}
