% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.ldaDMP.R
\name{print.ldaDMP}
\alias{print.ldaDMP}
\title{Printing object from \emph{print.ldaDMP} class by simple
print methods.}
\usage{
\method{print}{ldaDMP}(x, digits = getOption("digits"), ...)
}
\arguments{
\item{x}{Object from class \emph{glmDataSet} or from class
\emph{"RangedGlmDataSet"}.}

\item{digits}{Number of significant digits to be used.}
}
\description{
Printing object from \emph{print.ldaDMP} class by simple
print methods.
}
\keyword{internal}
