% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluateDIMPclass.R
\name{print.ldaDMP}
\alias{print.ldaDMP}
\alias{print.mlDMP}
\alias{print.qdaDMP}
\title{Printing object from \emph{ldaDMP} and \emph{qdaDMP} classes by
simple print methods}
\usage{
\method{print}{ldaDMP}(x, digits = getOption("digits"), ...)

\method{print}{qdaDMP}(x, digits = getOption("digits"), ...)
}
\arguments{
\item{x}{Object from class \strong{\emph{ldaDMP}} or \strong{\emph{qdaDMP}}.}

\item{digits}{Number of significant digits to be used.}
}
\description{
Objects from \emph{ldaDMP} and and \emph{qdaDMP} classes are
yielded by function \emph{evaluateDIMPclass}. These objects carries the
information of the corresponding machine-learning (ml) models used to
classify DMPs.
}
\details{
The definition of these classes makes less complex the downstream
analyses.
}
\keyword{internal}
