% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.qdaDMP.R
\name{print.qdaDMP}
\alias{print.qdaDMP}
\title{Printing object from \emph{print.qdaDMP} class by simple
print methods.}
\usage{
\method{print}{qdaDMP}(x, digits = getOption("digits"), ...)
}
\arguments{
\item{x}{Object from class \emph{glmDataSet} or from class
\emph{"RangedGlmDataSet"}.}

\item{digits}{Number of significant digits to be used.}
}
\description{
Printing object from \emph{print.qdaDMP} class by simple
print methods.
}
\keyword{internal}
