% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_rcounts.R
\docType{data}
\name{rcounts}
\alias{rcounts}
\title{Simulated dataset of read counts used in the examples}
\format{
rcounts is a list of \code{\link{GRanges-class}} objects, each
element carrying in their meta-columns the following variables:
\describe{
\item{mC}{methylation level from the reference sample}
\item{uC}{methylation level from the treatment sample}
}
where the column labeled 'mC' carries the number methylated cytosines and
column 'uC' the number of unmethylated cytosines.
The list carries 6 elements:
\describe{
\item{ref}{A reference sample to be used for the computation of
information divergences with function \code{\link{estimateDivergence}}.}
\item{C1 to C3}{GRanges from the control samples.}
\item{T1 to T3}{GRanges from the treatment samples.}
}
}
\usage{
rcounts
}
\description{
Each individuals sample includes 10000 cytosine positions
}
\keyword{datasets}
