% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shannonEntr.R
\name{shannonEntr}
\alias{shannonEntr}
\title{Compute Shannon Entropy}
\usage{
shannonEntr(p, logbase = 2)
}
\arguments{
\item{p}{A probability vector, sum(p) = 1.}

\item{logbase}{A positive number: the base with respect to which logarithms}
}
\description{
Compute Shannon Entropy of probability vector p.
}
\details{
By definition, if p_i = 0 for some i, the value of the corresponding
    summ and 0*log(0) is taken to be 0.
}
\examples{
counts = sample.int(10)
prob = counts/sum(counts)
shannonEntr(prob)
}
