% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simpleCutPoint.R
\name{simpleCutPoint}
\alias{simpleCutPoint}
\title{Simple Cutpoint estimation}
\usage{
simpleCutPoint(
  LR,
  control.names,
  treatment.names,
  column,
  div.col,
  tv.col = NULL,
  tv.cut,
  clas.perf,
  classifier,
  interactions = NULL,
  n.pc,
  prop,
  post.cut = 0.5,
  maxnodes = NULL,
  ntree = 400,
  nsplit = 1L,
  cutp_data = FALSE,
  num.cores,
  tasks,
  ...
)
}
\arguments{
\item{LR, res, control.names, treatment.names, column, div.col}{Same as in
\code{\link{estimateCutPoint}}}

\item{column, div.col, tv.col, tv.cut, clas.perf}{Same as in
\code{\link{estimateCutPoint}}}

\item{classifier, n.pc, prop, cutp_data, num.cores, tasks}{Same as in
\code{\link{estimateCutPoint}}}

\item{interactions}{If a logistic classifier is used. Variable interactions
to consider in a logistic regression model. Any pairwise combination of the
variable 'hdiv', 'TV', 'wprob', and 'pos' can be provided. For example:
'hdiv:TV', 'wprob:pos', 'wprob:TV', etc.}
}
\value{
Specified in function \code{\link{estimateCutPoint}} for parameter
setting \emph{simple = TRUE}
}
\description{
Internal function to estimate simple cutpoint according to
Youden Index.
}
\details{
This function is called by function
\code{\link{estimateCutPoint}}.
}
\examples{
## Get a set of potential DMPS (PS)
data(PS, package = 'MethylIT')

cutp <- simpleCutPoint(LR = PS, column = c(hdiv = TRUE, TV = TRUE,
                                           wprob = TRUE, pos = TRUE),
                       classifier = 'qda', n.pc = 4,
                       control.names = c('C1', 'C2', 'C3'),
                       treatment.names = c('T1', 'T2', 'T3'),
                       tv.cut = 0.5, clas.perf = TRUE, prop = 0.6,
                       div.col = 9L)

cutp

}
\keyword{internal}
