% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slidingGRegions.R
\name{slidingGRegions}
\alias{slidingGRegions}
\title{Generates intervals for a GRanges objects}
\usage{
slidingGRegions(
  GR,
  win.size = 1,
  step.size = 1,
  num.cores = 1L,
  tasks = 0,
  verbose = FALSE,
  ...
)
}
\arguments{
\item{GR}{A \code{\link[GenomicRanges]{GRanges-class}} object or a list
of \code{\link[GenomicRanges]{GRanges-class}} objects or an object that can
be coerced to a list of \code{\link[GenomicRanges]{GRanges-class}} objects.}

\item{win.size}{An integer. The size of the windows/intervals genomics.}

\item{step.size}{Interval at which the regions/windows must be defined}

\item{num.cores, tasks}{Parameters for parallel computation using package
\code{\link[BiocParallel]{BiocParallel-package}}: the number of cores to
use, i.e. at most how many child processes will be run simultaneously (see
\code{\link[BiocParallel]{bplapply}} and the number of tasks per job (only
for Linux OS).}

\item{verbose}{Logical. Default is TRUE. If TRUE, then the progress of the
computational tasks is given.}

\item{...}{Not in use.}
}
\value{
An object from \code{\link[IRanges]{IRangesList-class}}.
}
\description{
\strong{slidingGRegions} generates sliding intervals of a
specified width.
}
\details{
This function split a GRange object into intervals regions  of
fixed size. If win.size == step.size, then non-overlapping windows are
obtained. \strong{slidingGRegions} function generates sliding windows within
each range of IRL, according to width and step, returning a
\code{\link[IRanges]{IRangesList-class}}. If the sliding windows do not
exactly cover a range in IRL, the last window is partial.
}
\author{
Robersy Sanchez (\url{https://genomaths.com}).
}
