% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sortBySeqnameAndStart.R
\name{sortBySeqnameAndStart}
\alias{sortBySeqnameAndStart}
\alias{sortBySeqnameAndEnd}
\title{Sorting 'GRanges' objects}
\usage{
sortBySeqnameAndStart(gr)

sortBySeqnameAndEnd(gr)
}
\arguments{
\item{gr}{GRanges object}
}
\value{
GRanges object
}
\description{
Sorts a GRanges object by seqname and start position
}
\examples{
GR <- as(c('chr2:1-1', 'chr1:1-1'), 'GRanges')
GR <- sortBySeqnameAndStart(GR)

}
