% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uniqueGRanges.R
\name{uniqueGRanges}
\alias{uniqueGRanges}
\title{Unique genomic ranges from a list of GRanges objects}
\usage{
uniqueGRanges(
  ListOfGranges,
  ncols = NULL,
  columns = NULL,
  chromosomes = NULL,
  maxgap = -1L,
  minoverlap = 1L,
  missing = 0,
  type = c("any", "start", "end", "within", "equal"),
  select = c("all", "first", "last", "arbitrary"),
  ignore.strand = FALSE,
  keep.strand = !ignore.strand,
  num.cores = 1,
  tasks = 0L,
  verbose = TRUE
)
}
\arguments{
\item{ListOfGranges}{Objects to combine. A list of
\code{\link[GenomicRanges]{GRanges-class}} object or a
\code{\link[GenomicRanges]{GRangesList-class}} object.}

\item{ncols}{integer. Number of columns to use from the meta-column of each
GRanges object. Default value: NULL. If NULL, all the columns (from
column 1 to ncols) from each GRanges will be present in the
\eqn{uniqueGRanges} output.}

\item{columns}{integer number(s) corresponding to the specific column(s) to
use from the meta-column of each GRanges. Default value: NULL. if
provided, the metacolumn from the uniqueGRanges output will contain the
specified columns.}

\item{chromosomes}{Chromosomes used Default value: NULL}

\item{maxgap, minoverlap, ignore.strand, select, type}{The same as in
\code{\link[GenomicRanges]{findOverlaps-methods}}.}

\item{missing}{A numerical value (default 0) or NA to write in ranges with
missing values. For example, suppose that we want to build a
uniqueGRanges object from the GRanges objects X and Y. If a given range
k from the GRanges object X with metacolumn value x is missing in the
GRanges object Y, then the metacolumn of range k from
uniqueGRanges(list(X,Y)) object will be the row vector (x,0) or (x,NA)
if missing = NA.}

\item{keep.strand}{When set to TRUE, the strand information is preserved on
the objects even if ignore.strand is set to TRUE.  This makes it possible
to ignore the strand during overlap calculations but to preserve the
strand information and not overwrite with *. Default value is
keep.strand = !ignore.strand.}

\item{num.cores}{The number of cores to use, i.e. at most how many child
processes will be run simultaneously (see bplapply function from
BiocParallel package).}

\item{tasks}{integer(1). The number of tasks per job. value must be a scalar
integer >= 0L. In this documentation a job is defined as a single call
to a function, such as bplapply, bpmapply etc. A task is the division of
the X argument into chunks. When tasks == 0 (default), X is divided as
evenly as possible over the number of workers (see MulticoreParam from
BiocParallel package).}

\item{verbose}{if TRUE, prints the function log to stdout}
}
\value{
a GRanges object
}
\description{
Build an unique GRanges object from a list of Granges objects.
}
\details{
The metadata of each one of these GRanges must have one or more
columns to yield a unique GRanges object with metadata columns from the
original GRanges objects. Otherwise, a unique GRanges object will be
created  without metadata columns. Additionally, all metadata must be the
same class, e.g. all numeric or all characters, or all factor
}
\examples{
dfChr1 <- data.frame(chr = 'chr1', start = 11:15, end = 11:15,
                    strand = c('+','-','+','*','.'), score = 1:5)
dfChr2 <- data.frame(chr = 'chr1', start = 11:15, end = 11:15,
                    strand = c('+','-','+','*','.'), score = 1:5)
dfChr3 <- data.frame(chr = 'chr1', start = 11:15, end = 11:15,
                    strand = c('+','-','+','*','.'), score = 1:5)

gr1 <- makeGRangesFromDataFrame(dfChr1, keep.extra.columns = TRUE)
gr2 <- makeGRangesFromDataFrame(dfChr2, keep.extra.columns = TRUE)
gr3 <- makeGRangesFromDataFrame(dfChr3, keep.extra.columns = TRUE)

grList <- GRangesList('gr1' = gr1, 'gr2' = gr2, 'gr3' = gr3)

uniqueGRanges(grList)

}
\author{
Robersy Sanchez (\url{https://genomaths.com}).
}
