% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unlistX.R
\name{unlist.pDMP}
\alias{unlist.pDMP}
\alias{unlist.InfDiv}
\title{Flatten Lists extended to 'pDMP' and 'InfDiv' classes}
\usage{
unlist.pDMP(x)

unlist.InfDiv(x)
}
\arguments{
\item{x}{AN object from the class 'pDMP'.}
}
\description{
Given a 'pDMP' or 'InfDiv' objects, unlist simplifies it to
    produce a GRanges object which contains all the GRanges components which
    occur in 'pDMP' or 'InfDiv' object.
}
\details{
This is a method to extend unlist generic function to handle
    the 'pDMP' classes of objects
}
\examples{
gr1 <-GRanges(seqnames = "chr2", ranges = IRanges(3, 6),
          strand = "+", score = 5L, GC = 0.45)
gr2 <-
  GRanges(seqnames = c("chr1", "chr1"),
          ranges = IRanges(c(7,13), width = 3),
          strand = c("+", "-"), score = 3:4, GC = c(0.3, 0.5))
gr3 <-
  GRanges(seqnames = c("chr1", "chr2"),
          ranges = IRanges(c(1, 4), c(3, 9)),
          strand = c("-", "-"), score = c(6L, 2L), GC = c(0.4, 0.1))
grl <- list("gr1" = gr1, "gr2" = gr2, "gr3" = gr3)
# unlist(grl) # It does not work
class(grl) <-'InfDiv' # A trick
unlist(grl) # It works
}
