% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validateClass.R
\name{validateClass}
\alias{validateClass}
\alias{validateClass.default}
\alias{validateClass.pDMP}
\alias{validateClass.InfDiv}
\title{Function to validate S3 classes in MethylIT}
\usage{
validateClass(LR)

\method{validateClass}{default}(LR)

\method{validateClass}{pDMP}(LR)

\method{validateClass}{InfDiv}(LR)
}
\arguments{
\item{LR}{An object from class 'pDMP' or 'InfDiv'}
}
\description{
Function to validate S3 classes in MethylIT
}
\keyword{internal}
