% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimateBetaDist.R
\name{.estimateBetaDist}
\alias{.estimateBetaDist}
\title{Select the beta distribution that fit specified quantiles}
\usage{
.estimateBetaDist(q, init.pars = c(1, 1))
}
\arguments{
\item{q}{prior probabilities}

\item{init.pars}{initial parameter values. Defaults to alpha = 1 &
beta = 1, which imply the parsimony pseudo-counts greater than zero.}
}
\value{
the estimated values of the shape parameters of the selected beta
   distribution.
}
\description{
This function perform a rough estimation of the shape
    parameters of beta distribution
}
\details{
In order to obtain the estimates for shape parameters of beta
    distribution, the squared of the difference between the empirical
    cumulative distribution function (ecdf) & the theoretical cdf is 
    minimized using the Non-Linear Minimization function of 'stats' 
    package.
}
\keyword{internal}
