% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getDIMPatGenes.R
\name{getDIMPatGene}
\alias{getDIMPatGene}
\alias{getDIMPatGenes}
\alias{getDIMPatGenes.default}
\alias{getDIMPatGenes.GRanges}
\alias{getDIMPatGenes.pDMP}
\alias{getDIMPatGenes.InfDiv}
\alias{getDIMPatGenes.list}
\title{Count DIMPs at gene-body}
\usage{
getDIMPatGenes(GR, GENES, ignore.strand = TRUE)

\method{getDIMPatGenes}{default}(GR, GENES, ignore.strand = TRUE)

\method{getDIMPatGenes}{GRanges}(GR, GENES, ignore.strand = TRUE)

\method{getDIMPatGenes}{pDMP}(GR, GENES, ignore.strand = TRUE)

\method{getDIMPatGenes}{InfDiv}(GR, GENES, ignore.strand = TRUE)

\method{getDIMPatGenes}{list}(GR, GENES, ignore.strand = TRUE)
}
\arguments{
\item{GR}{An obects object from the any of the classes: 'pDMP', 'InfDiv',
GRangesList, GRanges or a list of GRanges.}

\item{GENES}{A GRanges object with gene coordinates and gene IDs. A
meta-column named 'gene_id' carying the gene ids should be included. If
the meta-column named 'gene_id' is not provided, then gene (region) ids
will be created using the gene (region) coordinates.}

\item{ignore.strand}{When set to TRUE, the strand information is ignored in
the calculations. Default value: TRUE}

\item{...}{Not in use.}
}
\value{
A GRanges object
}
\description{
The function counts DIMPs overlapping with gene-body. In fact,
    this function also can be used to count DIMPs overlapping with any set of
    regions given in a GRanges object.
}
\examples{
## Gene annotation
genes <- GRanges(seqnames = "1",
                ranges = IRanges(start = c(3631, 6788, 11649),
                                 end = c(5899, 9130, 13714)),
                strand = c("+", "-", "-"))
mcols(genes) <- data.frame(gene_id = c("AT1G01010", "AT1G01020",
                                       "AT1G01030"))

## Get a dataset of potential signals and the estimated cutpoint from the
## package
data(PS, cutpoint)

## The estimated cutpoints are used to discriminate signals from the noise.
## That is, DMPs are selected using the cupoints
DIMPs <- selectDIMP(PS, div.col = 9L, cutpoint = cutpoint$cutpoint)

## Finally DMPs found on genes
DIMR <- getDIMPatGenes(GR = DIMPs$T1, GENES = genes)

}
