% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_HD.R
\docType{data}
\name{HD}
\alias{HD}
\title{Simulated dataset of Hellinger divergences used in the examples}
\format{
'hdiv is an object from class 'InfDiv' with
\code{\link[GenomicRanges]{GRanges-class}} as elements carrying in the
meta-columns the following variables:
\describe{
\item{c1: }{Methylated read counts from the reference sample.}
\item{t1: }{Unmethylated read counts from the reference sample.}
\item{c2: }{Methylated read counts from the treatment sample.}
\item{t2: }{Unmethylated read counts from the treatment sample.}
\item{p1: }{methylation level from the reference sample.}
\item{p2: }{methylation level from the treatment sample.}
\item{TV: }{the total variation distance (difference of
methylation levels)}
\item{bay.TV: }{TV computes with Bayesian correction for the
methelation levels.}
\item{hdiv: }{Hellinger divergence.}
\item{jdiv: }{J divergence.}
}
'HD' was obtained with function \code{\link{estimateDivergence}}.
}
\usage{
HD
}
\description{
Each individuals sample includes 10000 cytosine positions
}
\examples{
## Load a dataset of simulated read counts.
data("HD", package = "MethylIT")

## Number of positions with Hellinger divergence values greater than
## the HD_95\%
critical.val <- do.call(rbind, lapply(HD, function(x) {
  hd.95 = quantile(x$hdiv, 0.95)
  tv.95 = quantile(abs(x$bay.TV), 0.95)
  return(c(tv = tv.95, hd = hd.95, num.sites.hd95 = sum(x$hdiv > hd.95),
           num.sites.tv95 = sum(x$bay.TV > tv.95)))}))
critical.val
}
\keyword{datasets}
