% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlCutpoint.R
\name{mlCutpoint}
\alias{mlCutpoint}
\title{Machine-learning Approach for Cutpoint Estimation}
\usage{
mlCutpoint(
  LR,
  control.names,
  treatment.names,
  column,
  div.col,
  tv.col = NULL,
  tv.cut,
  post.cut = 0.5,
  classifier1,
  classifier2 = NULL,
  interactions = NULL,
  n.pc,
  prop = 0.6,
  center = FALSE,
  scale = FALSE,
  stat = 0L,
  cut.values = NULL,
  maxnodes = NULL,
  ntree = 400,
  nsplit = 1L,
  num.cores = 1L,
  tasks = 0L,
  ...
)
}
\arguments{
\item{LR, res, control.names, treatment.names, column, div.col}{Same as in
\code{\link{estimateCutPoint}}}

\item{column, div.col, tv.col, tv.cut, cut.values, stat}{Same as in
\code{\link{estimateCutPoint}}}

\item{classifier1, classifier2, n.pc, prop, post.cut}{Same as in
\code{\link{estimateCutPoint}}}

\item{interactions}{If a logistic classifier is used. Variable interactions
to consider in a logistic regression model. Any pairwise combination of the
variable 'hdiv', 'TV', 'wprob', and 'pos' can be provided. For example:
'hdiv:TV', 'wprob:pos', 'wprob:TV', etc.}

\item{center}{A logical value indicating whether the variables should be
shifted to be zero centered.}

\item{scale}{A logical value indicating whether the variables should be}

\item{maxnodes, ntree}{Only for Random Forest classifier
(\code{\link[randomForest]{randomForest}}, 'random_forest'). Maximum number
\eqn{maxnodes} of terminal nodes trees in the forest can have. If not given,
trees are grown to the maximum possible. Parameter \eqn{ntree} stands for
the number of trees to grow. This should not be set to too small a number,
to ensure that every input row gets predicted at least a few times.}

\item{nsplit}{Only for Random Forest classifier. The Random Forest
(\code{\link[randomForest]{randomForest}}, 'random_forest') package uses a
C+Fortran implementation which only supports integer indexes, so any
dataframe/data table/matrix with >2^31 elements (limit for integers) gives
an error. The option nsplit is applied to train
\eqn{ntrees=floor(ntree/nsplit)} models with \eqn{rep(ntrees,nsplit)} which
are finally combined to obtain a forest with \eqn{ntree}. Each model in this
would contain \eqn{ntrees}.}

\item{...}{Additional arguments for \code{\link{evaluateDIMPclass}} function}

\item{cutp_data, num.cores, tasks}{Same as in \code{\link{estimateCutPoint}}}
}
\value{
Specified in function \code{\link{estimateCutPoint}} for parameter
setting \emph{simple = FALSE}
}
\description{
Internal function to estimate cutpoint following a
machine-learning approach
}
\details{
This function is called by function \code{\link{estimateCutPoint}}.
}
\examples{
## Get a set of potential DMPS (PS)
data(PS, package = 'MethylIT')

cutp <- mlCutpoint(LR = PS,
                 column = c(hdiv = TRUE, TV = TRUE,
                            wprob = TRUE, pos = TRUE),
                 classifier1 = 'qda', n.pc = 4,
                 control.names = c('C1', 'C2', 'C3'),
                 treatment.names = c('T1', 'T2', 'T3'),
                 tv.cut = 0.68, prop = 0.6,
                 div.col = 9L)
cutp$testSetPerformance
}
\author{
Robersy Sanchez (\url{https://genomaths.com}).
}
\keyword{internal}
