% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predictLogisticR.R
\name{predict.LogisticR}
\alias{predict.LogisticR}
\title{Predict function for 'LogisticR' method}
\usage{
\method{predict}{LogisticR}(object, newdata, type = c("class",
  "probabilities", "all"), ...)
}
\arguments{
\item{object}{To use with function 'predict'. A 'glm' object from a logistic
model containing a list of two objects: 1) an object of class inheriting
from "lda" and 2) an object of class inheriting from "prcomp".}

\item{newdata}{To use with function 'predict'. New data for classification
prediction}

\item{type}{To use with function 'predict'. . The type of prediction
required. The default is "all" given by function 'predict.lda' from MASS
package: 'class', 'posterior', and 'scores' (see ?predict.lda).}

\item{...}{Not in use.}
}
\value{
A character vector of prediction classes or a numeric vector of
  probabilities or a list containing the two vectors: prediction classes
  and probabilities.
}
\description{
Predict using a PCA-LDA model built with function 'LogisticR'
}
\details{
This function is specific for predictions based on a logistic model
    given by function 'evaluateDIMPclass'. A logistic model obtained with
    'glm' regression can be used directly with function 'predict' from
    'stats' package.
}
