% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predictDIMPclass.R
\name{predictDIMPclass}
\alias{predictDIMPclass}
\title{Predict DIMP class}
\usage{
predictDIMPclass(LR, model, conf.matrix = FALSE, control.names = NULL,
  treatment.names = NULL)
}
\arguments{
\item{LR}{A list of GRanges objects obtained through the through MethylIT
downstream analysis. Basically, this object is a list of GRanges containing
only differentially methylated position (DMPs). The metacolumn of each
GRanges must contain the columna: Hellinger divergence "hdiv", total
variation "TV", the probability of potential DMP "wprob", which naturally
are added in the downstream analysis of MethylIT.}

\item{model}{A classifier model obtained with the function
'evaluateDIMPclass'.}

\item{conf.matrix}{Optional. Logic, whether a confusion matrix should be
returned (default, FALSE, see below).}

\item{control.names}{Optional. Names/IDs of the control samples, which must
be include in thr variable LR (default, NULL).}

\item{treatment.names}{Optional. Names/IDs of the treatment samples, which
must be include in the variable LR (default, NULL).}
}
\value{
The same LR object with a column named "class" added to a GRanges
    object from LR (default). Based on the model prediction each DIMP is
    labeled as control "CT" or as treatment "TT". If "conf.matrix" is TRUE and
    the arguments control.names and treatment.names are provided, then the
    overall confusion matrix is returned
}
\description{
This function classify each DIMP as a control or a treatment
    DIMP
}
\details{
Predictions only makes sense if the query DIMPs belong to same
    methylation context and derive from an experiment accomplished under the
    same condition set for the DIMPs used to build the model.
}
\examples{
library(MethylIT)

data(cutpoint, PS, package = "MethylIT")

## DIMPs are selected using the cupoints
DMPs <- selectDIMP(PS, div.col = 9L, cutpoint = cutpoint$cutpoint,
                   tv.cut = 0.92)

## Classification of DIMPs into two clases: DIMPS from control and DIMPs from
## treatment samples and evaluation of the classifier performance (for more
## details see ?evaluateDIMPclass).
perf <- evaluateDIMPclass(LR = DMPs,
                          column = c(hdiv = TRUE, TV = TRUE,
                                     wprob = TRUE, pos = TRUE),
                          classifier = "lda", n.pc = 4L,
                          control.names =  c("C1", "C2", "C3"),
                          treatment.names = c("T1", "T2", "T3"),
                          center = TRUE, scale = TRUE,
                          prop = 0.6)

#' Now predictions of DIMP for control and treament can be obtained
pred = predictDIMPclass(LR = DMPs, model = perf$model,
                        conf.matrix = TRUE,
                        control.names = c("C1", "C2", "C3"),
                        treatment.names = c("T1", "T2", "T3"))
}
