% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glmDataSet.R
\name{print.glmDataSet}
\alias{print.glmDataSet}
\alias{print.RangedGlmDataSet}
\title{Printing object from \strong{\emph{glmDataSet}} and
\strong{\emph{'RangedGlmDataSet'}} classes by simple print methods.}
\usage{
\method{print}{glmDataSet}(x, digits = getOption("digits"), ...)

\method{print}{RangedGlmDataSet}(x, digits = getOption("digits"), ...)
}
\arguments{
\item{x}{Object from class \emph{glmDataSet} or from class
\emph{'RangedGlmDataSet'}.}

\item{digits}{Number of significant digits to be used.}
}
\description{
Printing object from \strong{\emph{glmDataSet}} and
\strong{\emph{'RangedGlmDataSet'}} classes by simple print methods.
}
\keyword{internal}
