% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validateClass.R
\name{validateClass}
\alias{validateClass}
\alias{validateClass.default}
\alias{validateClass.pDMP}
\alias{validateClass.InfDiv}
\title{Function to validate S3 classes in MethylIT}
\usage{
validateClass(LR)

\method{validateClass}{default}(LR)

\method{validateClass}{pDMP}(LR)

\method{validateClass}{InfDiv}(LR)
}
\arguments{
\item{LR}{An object from class 'pDMP' or 'InfDiv'}
}
\value{
TRUE if object is valid, otherwise warning() and stop() are called
}
\description{
S3 generic used internally by MethylIT to ensure correct object
classes are being used.  validateClass() is called by these functions:
estimateCutPoint evaluateDIMPclass FisherTest getPotentialDIMP
nonlinearFitDist predictDIMPclass predictLogisticR selectDIMP
uniqueGRfilterByCov
}
\examples{
data(HD, gof)
PS <- getPotentialDIMP(LR = HD, nlms = gof$nlms, div.col = 9L, alpha = 0.05)
validateClass(HD)
validateClass(PS)

}
\keyword{internal}
